Rename main layout component, Add `focus-visible` to storybook in `preview.js`, Add main layout story demo component for storybook support,
