+++ src/04_templates/MainLayout.stories.js
@@ -1,0 +1,9 @@
+import React from "react"
+import Layout from "../04_templates/MainLayout-story"
+
+export default {
+    title: "/04 Templates/Main Layout",
+    component: Layout,
+}
+
+export const Working = () => <Layout />
