+++ src/04_templates/PageTemplate.js
@@ -5,7 +5,7 @@
 
 import React from "react"
 import { graphql } from "gatsby"
-import Layout from "./LayoutTemplate"
+import Layout from "./MainLayout"
 import SEO from "../09_utilities/SEO"
 
 export default function Template({
