+++ src/04_templates/PostTemplate.js
@@ -6,7 +6,7 @@
 import React from "react"
 import { graphql } from "gatsby"
 import { Link } from "gatsby"
-import Layout from "./LayoutTemplate"
+import Layout from "./MainLayout"
 import SEO from "../09_utilities/SEO"
 
 export default function Template({
