+++ src/pages/markdown.js
@@ -6,7 +6,7 @@
 import React from "react"
 import { Link } from "gatsby"
 import SEO from "../09_utilities/SEO"
-import Layout from "../04_templates/LayoutTemplate"
+import Layout from "../04_templates/MainLayout"
 
 const MarkdownPage = () => (
     <>
